refactored getQuizzes functions in all tabs into one custom hook, useQuizzes
