+++ client/src/components/user-dashboard/components/Favorites.jsx
@@ -1,28 +1,16 @@
 import { Typography, TableBody, TableCell, TableHead, TableRow, Table, Stack } from '@mui/material';
 import FilterBar from './subComponents/FilterBar.jsx';
 import FavoriteIcon from '@mui/icons-material/Favorite';
-import { useState, useEffect } from 'react';
+import useQuizzes from '../hooks/useQuizzes';
 
 const Favorites = (props) => {
 
-  const [quizzes, setQuizzes] = useState([]);
-
-  const getQuizzes = async (userId) => {
-    const url = process.env.REACT_APP_API_URI;
-    const response = await fetch(`${url}/dashboard/favorites`);
-    if (response.ok) {
-      setQuizzes(await response.json());
-    }
-  };
+  const quizzes = useQuizzes('favorites');
 
   const handleUnlike = (e) => {
     // handle unliking quiz
     console.log('unlike')
   };
-
-  useEffect(() => {
-    getQuizzes();
-  }, []);
 
   return (
     <>
